requirements.txt
torch
torchvision
lpips
Pillow